+++ packages/eval-suite/src/index.ts
@@ -2,6 +2,7 @@ export * from "./types.js";
 export * from "./dogfoodEvalCases.js";
 export * from "./enterpriseDogfoodEvalCases.js";
 export * from "./evalCases.js";
+export * from "./redTeamEvalCases.js";
 export * from "./runEvalCase.js";
 export * from "./runEvalSuite.js";
 export * from "./summarizeEvalResults.js";